// src/server/model/iItemGroup.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { IAttributeGroup } from './iAttributeGroup';


export interface IItemGroup { 
    hsnCode?: string;
    parentItemGroupId?: number;
    parentItemGroupName?: string;
    filterAttributeGroupId?: number;
    filterAttributeGroupName?: string;
    featureAttributeGroups?: Array<IAttributeGroup>;
    systemGenerated?: boolean;
    name?: string;
    description?: string;
    importedId?: number;
    logInfo?: string;
    productUserId?: number;
    creationdate?: Date;
    productUserName?: string;
    modificationdate?: Date;
    impCompanyGuid?: string;
    id?: number;
    jacksontype?: string;
}